# swagger-ui/swagger.json
{
  "swagger": "2.0",
  "paths": {
    "/api/v1/find_mentors/{topic}/{level}": {
      "get": {
        "description": "Returns a list of mentors that are listed for the given topic at or above the given level",
        "tags": [
          "Mentors"
        ],
        "operationId": "FindMentors",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "TopicName",
            "description": "the name of the desired topic area",
            "name": "topic",
            "in": "path",
            "required": true
          },
          {
            "pattern": "[1-5]",
            "type": "integer",
            "format": "int64",
            "x-go-name": "Level",
            "description": "the desired level of proficiency in the desired topic",
            "name": "level",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/models.User"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/api/v1/mentor_request": {
      "post": {
        "description": "Constructs and persists a new MentorRequest based on the provided parameters",
        "tags": [
          "Mentors"
        ],
        "operationId": "NewMentorRequest",
        "parameters": [
          {
            "x-go-name": "MentorID",
            "description": "the ID of the mentor user",
            "name": "mentor_id",
            "in": "body",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "uniqueItems": true,
            "x-go-name": "MenteeID",
            "description": "the ID of the mentee user",
            "name": "mentee_id",
            "in": "body",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/models.MentorRequest"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/api/v1/mentor_topic": {
      "post": {
        "description": "Constructs and persists a new MentorTopic based on the provided parameters",
        "tags": [
          "Mentors"
        ],
        "operationId": "NewMentorTopic",
        "parameters": [
          {
            "x-go-name": "UserID",
            "description": "the ID of the mentor/user",
            "name": "user_id",
            "in": "body",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "x-go-name": "TopicID",
            "description": "the ID of the topic",
            "name": "topic_id",
            "in": "body",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "pattern": "[1-5]",
            "x-go-name": "Level",
            "description": "the self-assessed level of the mentor's skill in the topic",
            "name": "level",
            "in": "body",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "x-go-name": "Description",
            "description": "a brief description of the mentor's experience in the topic",
            "name": "description",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/models.MentorTopic"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/api/v1/topic": {
      "post": {
        "description": "Constructs and persists a new Topic based on the provided parameters",
        "tags": [
          "Topics"
        ],
        "operationId": "NewTopic",
        "parameters": [
          {
            "uniqueItems": true,
            "x-go-name": "Name",
            "description": "the name of the new topic",
            "name": "name",
            "in": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/models.Topic"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/api/v1/topic/{id}": {
      "get": {
        "tags": [
          "Topics"
        ],
        "summary": "Gets info for a topic.",
        "operationId": "GetTopic",
        "parameters": [
          {
            "uniqueItems": true,
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "the ID of the topic",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/models.Topic"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/api/v1/topic_like/{part}": {
      "get": {
        "description": "Gets topics similar to the given string",
        "tags": [
          "Topics"
        ],
        "operationId": "GetTopicsLike",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Part",
            "description": "string to search for similar topics",
            "name": "part",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/models.Topic"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/api/v1/user": {
      "post": {
        "description": "Constructs and persists a new user based on the provided parameters",
        "tags": [
          "Users"
        ],
        "operationId": "NewUser",
        "parameters": [
          {
            "uniqueItems": true,
            "x-go-name": "Username",
            "description": "the username of the new user",
            "name": "username",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "x-go-name": "DisplayName",
            "description": "the pretty name or nickname of the new user",
            "name": "display_name",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "uniqueItems": true,
            "x-go-name": "Email",
            "description": "the email of the new user",
            "name": "email",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "x-go-name": "Password",
            "description": "the password of the new user",
            "name": "password",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "x-go-name": "Description",
            "description": "a brief self-description of the user",
            "name": "description",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "x-go-name": "IconURL",
            "description": "a URL pointing to the user's icon",
            "name": "icon_url",
            "in": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/models.User"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
    "/api/v1/user/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Gets info for a user.",
        "operationId": "GetUser",
        "parameters": [
          {
            "uniqueItems": true,
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "the ID of the user",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/models.User"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    }
  },
  "definitions": {}
}